+++ testData/folding/DifferentFunctions.bash
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+function a() {
+    X="def value"
+}
+
+function b() {
+    "$X"
+}