+++ Changelog.md
@@ -1,6 +1,12 @@
+#### 2018-08-25:
+ - fix to improve handling of local variables in functions
+
 #### 2018-08-14:
  - new icons for plugin (contributed by nosovae-dev)
- 
+
+#### 2018-08-09:
+ - Feature: Add folding for variables (contributed by nosovae-dev)
+
 #### 2018-08-09:
  - Add kotlin language for development plugin (contributed by nosovae-dev)